requirements.txt
Django
django-cors-headers 
django-environ 
djangorestframework 
requests
